Add a basic HTML view (HTMLPage class)
The HTML view provides facilities for creating HTML documents in
memory, which can then be served to the client as web pages.

This commit also provides an initial implementation of a HTML view for
track metadata. This view is served when a client issues a GET request
to /tracks/, and lists information about each public track in the
track database.