html-view/track-metadata: Remove Loader version field
Removes the field that provided the required version number of the
RallySportED Loader to play the track.

I'm of two minds about including this information in the metadata view.
On the one hand, the fewer fields the view has, the easier it is to
read; but on the other hand, it's useful for the user to know in advance
which version of Loader the track requires - the Loader isn't always
backwards-compatible, for instance.

Right now, I'm erring on the side of not including this information, as
I think the metadata view was otherwise getting a bit too info-heavy.
This may mean the user has to do a bit of trial-and-error (download the
track, try to run it, be informed that their current Loader is not
supported), which isn't ideal. But they can also always convert the track
into a format supported by the most recent version of the Loader by
opening it in RallySportED-js and exporting it back out.

That said, I'll be considering ways to include this information in the
view that don't contribute as much to information overload.