Re-license from GPL3 to GPL2-or-later
A recent source file, zip-file.php, comes with some uncertainty about
whether it's under GPL2-only or GPL2-or-later.

So for now, the software as a whole is re-licensed under GPL2-or-later,
and when including the current incarnation of zip-file.php, distribution
is under GPL2-only.

Future versions of the program may replace zip-file.php with a different,
GPL2-or-later implementation.