track-database: Reduce amount of data stored in database
Instead of storing of each track's data a pre-generated downloadable zip
file (~50 KB) as well as a raw JSON file (~300 KB), we'll just store the
track container and manifesto files in a compressed format (~50 KB and
< 1 KB, respectively) and use them to dynamically generate the zip and
JSON files when requested.

In effect, this exchanges speed of serving (pre-generated content) for a
smaller database - although, since the data are pretty small anyway, the
effect on performance is probably not large either.