html-view/view-track: Display a random selection of 30 tracks
By default, the track view will now display 30 randomly-selected tracks.

This is a bit of a lazy kludge for now - it limits the maximum number of
tracks shown at once (so the page doesn't get overly long when the
database contains a lot of tracks), while obviating the need to implement
pagination to navigate through all the tracks.

Some kind of pagination will likely be implemented in the future, though.