html-view/form: Introduce inner_html()
While html() provides the form's entire HTML code, inner_html() is the
code for the <form> element only, excluding the container <div> etc.
This way, derived forms can override inner_html() rather than the entire
html(), and thus not have to bother with setting up the container and
so on.