Require registration track file in order to reset user password
A user wishing to request a password reset must now provide the track
file they registered with, as an access key of sorts. This will
hopefully prevent spamming of the reset request form.

If the user is unable to provide the track file, they are instructed
to contact us directly.